First optimization for #5461, which covers 2 fixes:
1. Regression from d8371bd

In **UsersController**, fix sent to query off the request table instead of the large list of packages. Performance was hard to measure, as it was hanging due to the regression.

2. Optimize packages query to only sort in DB, based on IsLatest columns

In **PackageService**, modified the EF query to sort and select based on IsLatest state in DB. Previously we'd read all rows and sort in memory (on NuGetVersion as last resort) to handle cases where IsLatest wasn't set due to concurrency issues.

The second optimization helps mostly in scenarios with many versions, as opposed to many registrations. Results below:

**Scenario 1:  59K registrations, 59K versions (NuGetStagingTest on DEV)**

  - BEFORE: 48.54, 49.41, 60, 35.74, 49.13 => 45.71s

  - AFTER: 36.12, 49.95, 43.13, 40.32, 46.30 => 41.47 **(-9.28%)**

**Scenario 2: 100 registrations, 52K versions (DEV, various packages)**

  - BEFORE: 126, 144, 90, 108, 78 => 100.5s

  - AFTER: 2.71, 3.15, 2.28, 2.23, 2.26 => 2.37s **(-97.64%)**