@chenriksson 2018-03-05
@joelverhagen Can you review my last push (663bceb)? I had a bug where I wasn't getting the correct latest version... EF doesn't preserve the ordering done before groupby like linq-to-objects does.

Also added a fix for #5375: `.ThenByDescending(p => p.Listed)` should order published versions first. UsersController.Packages applies filter `.Where(p => p.Listed)` which should exclude these from the Published list. (I'm assuming `!Listed` when `PackageState != Available`... but correct me if I'm wrong and I can also look at `PackageState`)

**Scenario 2: 100 registrations, 52K versions (re-run morning of 3/5)**
- BEFORE: 72s, 84s, 72s, 84s, 84s => 78s
- AFTER: 4.6s, 4.64s, 4.52s, 4.55s, 4.48s => 4.54s (-94.18%)