@joelverhagen 2018-03-05
> EF doesn't preserve the ordering done before groupby like linq-to-objects does.

WOW! That's a serious gotcha.

> `.ThenByDescending(p => p.Listed)`

This doesn't seem 100% critical since this is a package owner view, but seems fine.

> I'm assuming `!Listed` when `PackageState != Available`... 

Invalid assumption. You need to filter out non-Available if you actually want that.